8_computacao_conteiners_serverless/lab_lambda_ecr_s3/requirements.txt
Pillow==11.3.0
boto3==1.34.162
